move dataviews and picker scroll handlers and callback into a hook
