+++ packages/dataviews/src/hooks/index.ts
@@ -3,4 +3,5 @@
  */
 export { default as useFormValidity } from './use-form-validity';
 export { useData } from './use-data';
+export { useInfiniteScroll } from './use-infinite-scroll';
 export { default as useSelectedItems } from './use-selected-items';
